requirements.txt
gevent-websocket
Flask-Cors
flask-socketio
PySerial
